Added --acounts option to edit accounts without running applet (Ken Hayber) New translations: ru, uk (Konstantin Korikov) Get local mailbox dir from MAIL env variable if not otherwise defined (Konstantin Korikov) Change polltime from int to float ((Konstantin Korikov) Improved update-po script ((Konstantin Korikov)

+++ checker.py
@@ -9,10 +9,10 @@ class Checker:
 	def __init__(self, config):
 		try:
 			self.name = config['name']
-			self.polltime = int(config['polltime'])
+			self.polltime = float(config['polltime'])
 		except:
 			self.name = ''
-			self.polltime = 10
+			self.polltime = 10.0
 
 		self.blocker = None
 		self.results = ""
